fix: declare vu32 and vs32 at the width their names state
long is 64 bits on LP64 hosts, so volatile unsigned long and volatile long give
vu32 and vs32 a 64-bit width on every target this builds for. int is 32 bits
there. On the N64, where long is 32 bits, the two spellings agree.

Three sites read these types, all in N64 hardware paths the host build excludes:
src/lib/ultra/os/syncputchars.c casts a packet through vu32, and include/PR/rcp.h's
IO_READ and IO_WRITE cast an MMIO address through it. vs32 has no uses.

The compiled source set, the warning profile, and the ROM run are unchanged: 368
translation units, 0 errors, 177 warnings.

+++ include/PR/ultratypes.h
@@ -46,12 +46,12 @@ typedef signed long long int    s64; /* signed 64-bit */
 
 typedef volatile unsigned char		vu8;	/* unsigned  8-bit */
 typedef volatile unsigned short		vu16;	/* unsigned 16-bit */
-typedef volatile unsigned long		vu32;	/* unsigned 32-bit */
+typedef volatile unsigned int		vu32;	/* unsigned 32-bit */
 typedef volatile unsigned long long	vu64;	/* unsigned 64-bit */
 
 typedef volatile signed char		vs8;	/* signed  8-bit */
 typedef volatile short			vs16;	/* signed 16-bit */
-typedef volatile long			vs32;	/* signed 32-bit */
+typedef volatile signed int		vs32;	/* signed 32-bit */
 typedef volatile long long		vs64;	/* signed 64-bit */
 
 typedef float				f32;	/* single prec floating point */
